Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/scheme-others.sub.any-expected.txt

CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load aaa://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load cap://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load cid://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load dav://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load dict://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load dns://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load geo://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load im://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load imap://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load ipp://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load ldap://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load mailto://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load nfs://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load pop://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load rtsp://localhost:8800/ due to access control checks.
CONSOLE MESSAGE: line 7: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 7: Fetch API cannot load snmp://localhost:8800/ due to access control checks.

PASS Fetching aaa://localhost:8800/ is KO 
PASS Fetching cap://localhost:8800/ is KO 
PASS Fetching cid://localhost:8800/ is KO 
PASS Fetching dav://localhost:8800/ is KO 
PASS Fetching dict://localhost:8800/ is KO 
PASS Fetching dns://localhost:8800/ is KO 
PASS Fetching geo://localhost:8800/ is KO 
PASS Fetching im://localhost:8800/ is KO 
PASS Fetching imap://localhost:8800/ is KO 
PASS Fetching ipp://localhost:8800/ is KO 
PASS Fetching ldap://localhost:8800/ is KO 
PASS Fetching mailto://localhost:8800/ is KO 
PASS Fetching nfs://localhost:8800/ is KO 
PASS Fetching pop://localhost:8800/ is KO 
PASS Fetching rtsp://localhost:8800/ is KO 
PASS Fetching snmp://localhost:8800/ is KO 
